requirements.txt
streamlit>=1.28.0
google-cloud-texttospeech>=2.27.0
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
google-cloud-speech>=2.33.0
pyaudio>=0.2.11
japanese-address-parser-py>=0.1.24